terminando permissões

+++ ecommerce/resources/views/produtos/index.blade.php
@@ -10,9 +10,9 @@
         <div class="pull-right">
 
 
-
-            <a class="btn btn-success" href="{{ route('produtos.create') }}"> + Novo Produto</a>
-
+            @can('produtos-create')
+                <a class="btn btn-success" href="{{ route('produtos.create') }}"> + Novo Produto</a>
+            @endcan
 
 
         </div>
@@ -50,18 +50,21 @@
        <a class="btn btn-info" href="{{ route('produtos.show',$produto->id) }}">Mostrar</a>
 
 
-
-        <a class="btn btn-primary" href="{{ route('produtos.edit',$produto->id) }}">Editar</a>
-
-
-
+        @can('produtos-edit')
+            <a class="btn btn-primary" href="{{ route('produtos.edit',$produto->id) }}">Editar</a>
+        @endcan
 
 
-        {!! Form::open(['method' => 'DELETE','route' => ['produtos.destroy', $produto->id],'style'=>'display:inline']) !!}
+
+
+        @can('produtos-delete')
+            {!! Form::open(['method' => 'DELETE','route' => ['produtos.destroy', $produto->id],'style'=>'display:inline']) !!}
 
             {!! Form::submit('Apagar', ['class' => 'btn btn-danger']) !!}
 
-        {!! Form::close() !!}
+            {!! Form::close() !!}
+        @endcan
+
 
 
     </td>
